requirements.txt
alembic>=1.14,<2.0
asyncpg>=0.30,<1.0
cachetools>=5.5,<6.0
discord.py>=2.5,<3.0
pillow>=12.1,<13.0
pymongo>=4.11,<5.0
pydantic>=2.11,<3.0
pydantic-settings>=2.8,<3.0
python-dotenv>=1.0,<2.0
sqlalchemy>=2.0.39,<3.0
structlog>=25.1,<26.0
